+++ Casks/integra.rb
@@ -1,6 +1,6 @@
 cask "integra" do
-  version "0.15.4"
-  sha256 "fda5bbfee7fb0017afa5347dfd82237882291817bfc5621076788582ed429bd2"
+  version "0.16.0"
+  sha256 "0d4ee2b3b51b7d61978ac9778a07837a3b9486151dbfa3497a8c66e518bef536"
 
   url "https://github.com/Octadira/integra/releases/download/v#{version}/Integra-v#{version}.dmg"
   name "Integra"